refactor: make args of api client more readable and minor changes
